pyproject.toml
[tool.poetry]
name = "bl-matchmaker"
version = "0.1.0"
description = "Matchmaker service for Bannerlord Matchmaking system"
authors = ["Forsee41 <forsee41@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.88.0"
pydantic = "^1.10.2"
pyyaml = "^6.0"
uvicorn = "^0.20.0"
fastapi-utils = "^0.2.1"
httpx = "^0.23.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
mypy = "^0.991"
debugpy = "^1.6.4"

[tool.poe.tasks]
test = "python -m pytest -v"
run = "python app/main.py"

[tool.pytest.ini_options]
pythonpath = ["app"]

[tool.pyright]
executionEnvironments = [{ root = "tests", extraPaths = ["app"]}]

[tool.black]
line-length = 88

[mypy]
mypy_path = "app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


